test: add test transformToDispatchModel
